+++ setup.php
@@ -19,7 +19,7 @@ use GlpiPlugin\Uxcustomizer\ImpactMapTab;
 use GlpiPlugin\Uxcustomizer\Menu;
 use GlpiPlugin\Uxcustomizer\MenuOrder;
 
-define('PLUGIN_UXCUSTOMIZER_VERSION',          '2.0.1');
+define('PLUGIN_UXCUSTOMIZER_VERSION',          '2.1.0');
 define('PLUGIN_UXCUSTOMIZER_MIN_GLPI_VERSION', '11.0.0');
 define('PLUGIN_UXCUSTOMIZER_MAX_GLPI_VERSION', '11.0.99');